Moved PointReferenceTypes to their own files
