+++ OG/OG/ASTBuilding/TreeNodes/DeclarationNodes/ShapePointReference.cs
@@ -1,0 +1,13 @@
+﻿using OG.ASTBuilding.Terminals;
+
+namespace OG.ASTBuilding.TreeNodes.DeclarationNodes
+{
+    public abstract class ShapePointReference : PointReferenceNode
+    {
+        public ShapePointReference(string pointText, IdNode shapeName, PointReferenceNodeType type):base(pointText, type)
+        {
+            ShapeName = shapeName;
+        }
+        public IdNode ShapeName { get; set; }
+    }
+}
