+++ OG/OG/ASTBuilding/TreeNodes/DeclarationNodes/ShapeStartPointNode.cs
@@ -1,0 +1,11 @@
+﻿using OG.ASTBuilding.Terminals;
+
+namespace OG.ASTBuilding.TreeNodes.DeclarationNodes
+{
+    public class ShapeStartPointNode : ShapePointReference
+    {
+        public ShapeStartPointNode(string pointText, IdNode shapeName) : base(pointText,  shapeName, PointReferenceNodeType.ShapeStartPointNode)
+        {
+        }
+    }
+}
